requirements.txt
Django==4.1.5
django-bootstrap-v5==1.0.11
django-filter==22.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
drf-nested-routers==0.93.4
Pillow==9.4.0
pytils==0.4.1
mock==5.0.1
psycopg2==2.9.5
uwsgi==2.0.20








